Challenges/Web/compressor/src/requirements.txt
Flask==3.0.3
waitress==2.1.2
PyYAML==6.0.1